Fix all of ustrip, uconvert etc at the "root" level
The problem:
Running `ustrip.(ENU(1.0m, 2.0m, 3.0))` by default yields `SVector(1., 2., 3.)` not `ENU(1., 2., 3.)`.
I started overwriting `ustrip` etc manually, as well as `map`, but that
can't be the solution...
So I dug into the depths of `StaticArrays`, and at the core the
`copy(instantiate(broadcast))` gets a "default type" to copy to (when
broadcasting).
So specifying the default type for the types we want this to be true for
basically fixes everthing(!)